Participated in ISEA workshop organized by IIT Guwahati.
This workshop was organized by IIT Guwahati at SRM Institute of Science and Technology which was based upon Information Security Research Trends and Awareness. Information security awareness is an evolving part of information security that focuses on raising consciousness regarding potential risks of the rapidly evolving forms of information and the rapidly evolving threats.